feat(user): 新增用户文章和任务相关功能
- 新增 UserArticle 和 UserCreateTask 实体类
- 新增用户文章和任务相关 mapper 和 service
- 修改用户信息，使用 Long 类型表示完成任务数
- 优化文章和任务的 Markdown 处理逻辑- 调整用户登录和注册相关代码

+++ studyParty_entityApi/src/main/java/com/studyParty/entity/user/User.java
@@ -38,7 +38,7 @@ public class User implements Serializable {
     private String email;
     private Date lastLogin;
     private Date createDate;
-    private Integer finishTask;
+    private Long finishTask;
 
     public User(String name, String password) {
         this.name = name;
@@ -60,7 +60,7 @@ public class User implements Serializable {
         this.starPrestige = 0;
         this.clockIn = 1;
         this.lastLogin = Date.valueOf(LocalDate.now());
-        this.finishTask = 0;
+        this.finishTask = 0L;
     }
     /// 修改个人信息专用
     public User(Long id, String name, String sex, String major, String grade, String phone, String school, String email) {
